Module pathfinding::directed

Algorithms for directed graphs.

Modules

astar

Compute a shortest path (or all shorted paths) using the A* search algorithm.

bfs

Compute a shortest path using the breadth-first search algorithm.

count_paths

Count the total number of possible paths to reach a destination.

cycle_detection

Identify a cycle in an infinite sequence.

dfs

Compute a path using the depth-first search algorithm.

dijkstra

Compute a shortest path using the Dijkstra search algorithm.

edmonds_karp

Compute the maximum flow that can go through a directed graph using the Edmonds Karp algorithm.

fringe

Compute a shortest path using the Fringe search algorithm.

idastar

Compute a shortest path using the IDA* search algorithm.

iddfs

Compute a shortest path using the iterative deepening depth-first search algorithm.

strongly_connected_components

Separate nodes of a directed graph into strongly connected components.

topological_sort

Find a topological order in a directed graph if one exists.

yen

Compute k-shortest paths using Yen’s search algorithm.