Module pathfinding::directed::bfs

Compute a shortest path using the breadth-first search algorithm.

Structs

BfsReachable

Struct returned by bfs_reach.

Functions

bfs

Compute a shortest path using the breadth-first search algorithm.

bfs_loop

Return one of the shortest loop from start to start if it exists, None otherwise.

bfs_reach

Visit all nodes that are reachable from a start node. The node will be visited in BFS order, starting from the start node and following the order returned by the successors function.