Module pathfinding::directed::dijkstra

Compute a shortest path using the Dijkstra search algorithm.

Functions

build_path

Build a path leading to a target according to a parents map, which must contain no loop. This function can be used after dijkstra_all or dijkstra_partial to build a path from a starting point to a reachable target.

dijkstra

Compute a shortest path using the Dijkstra search algorithm.

dijkstra_all

Determine all reachable nodes from a starting point as well as the minimum cost to reach them and a possible optimal parent node using the Dijkstra search algorithm.

dijkstra_partial

Determine some reachable nodes from a starting point as well as the minimum cost to reach them and a possible optimal parent node using the Dijkstra search algorithm.