Function pathfinding::directed::dijkstra::dijkstra_reach

pub fn dijkstra_reach<N, C, FN, IN>(
    start: &N,
    successors: FN
) -> DijkstraReachable<N, C, FN> 
where
    N: Eq + Hash + Clone,
    C: Zero + Ord + Copy,
    FN: FnMut(&N, C) -> IN,
    IN: IntoIterator<Item = (N, C)>,

Visit all nodes that are reachable from a start node. The node will be visited in order of cost, with the closest nodes first.

The successors function receives the current node and the best cost up to this node, and returns an iterator of successors associated with their move cost.