Module pathfinding::directed::strongly_connected_components

Separate nodes of a directed graph into strongly connected components.

A path-based strong component algorithm is used.

Functions

strongly_connected_component

Compute the strongly connected component containing a given node.

strongly_connected_components

Partition all strongly connected components in a graph.

strongly_connected_components_from

Partition nodes reachable from a starting point into strongly connected components.