Module pathfinding::directed::topological_sort

Find a topological order in a directed graph if one exists.

Functions

topological_sort

Find a topological order in a directed graph if one exists.

topological_sort_into_groups

Topologically sort a directed graph into groups of independent nodes.