Module pathfinding::matrix

Matrix of an arbitrary type and utilities to rotate, transpose, etc.

Modules

directions

Directions usable for Matrix::in_direction() second argument.

Structs

ColumnIterator

Column iterator returned by column_iter() on a matrix.

Matrix

Matrix of an arbitrary type. Data are stored consecutively in memory, by rows. Raw data can be accessed using as_ref() or as_mut().

RowIterator

Row iterator returned by iter() on a matrix.

Enums

MatrixFormatError

Format error encountered while attempting to build a Matrix.